LayoutTests/media/media-can-play-type-expected.txt

Test HTMLMediaElement canPlayType() method.

EXPECTED (video.canPlayType() == '') OK
EXPECTED (video.canPlayType('video/') == '') OK
EXPECTED (video.canPlayType('video/blahblah') == '') OK
EXPECTED (video.canPlayType('video/blahblah; codecs="blah, , blah"') == '') OK
EXPECTED (video.canPlayType('video/blahblah; codecs=blah') == '') OK
EXPECTED (video.canPlayType('video/blahblah; codecs="blah"') == '') OK
EXPECTED (video.canPlayType('video/blahblah; codecs="badcontent') == '') OK
EXPECTED (video.canPlayType('video/blahblah; codecs=badcontent"') == '') OK
EXPECTED (video.canPlayType('video/blahblah; codecs="badcontent"') == '') OK
END OF TEST
